feat: Add half stars support to Rating component
