+++ features/collection/components/sections/FilterableProductsSection/ProductList.tsx
@@ -1,20 +1,18 @@
 import {
    AspectRatio,
    Badge,
-   Box,
    Button,
    Divider,
    Heading,
    HStack,
    Img,
-   StackProps,
    Text,
    VStack,
 } from '@chakra-ui/react';
+import { Rating } from '@components/Rating';
 import { ShopifyImage } from '@components/ShopifyImage';
 import { ProductHit } from '@features/collection';
 import * as React from 'react';
-import { FaStar } from 'react-icons/fa';
 
 const placeholderImageUrl =
    'https://via.placeholder.com/180x135?text=not+available';
@@ -170,26 +168,3 @@ export function ProductListItem({ product }: ProductListItemProps) {
       </HStack>
    );
 }
-
-interface RatingProps extends StackProps {
-   value?: number;
-}
-
-export const Rating = (props: RatingProps) => {
-   const { value = 5, ...rest } = props;
-   return (
-      <HStack spacing="1" {...rest}>
-         {Array.from({ length: 5 }).map((_, i) => {
-            const fade = i + 1 > value;
-            return (
-               <Box
-                  as={FaStar}
-                  color={fade ? 'gray.300' : 'blue.500'}
-                  fontSize="larger"
-                  key={i}
-               />
-            );
-         })}
-      </HStack>
-   );
-};
